backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.42.0
python-docx==1.1.2
pdfminer.six==20231228
Pillow==10.4.0
pydantic==2.8.2
djangorestframework-simplejwt==5.3.1

